+++ mex/editor/locale_service.py
@@ -3,16 +3,15 @@ from collections.abc import Iterable
 from gettext import GNUTranslations
 from importlib.resources import files
 from io import BytesIO
-from typing import TYPE_CHECKING, Self, cast
+from pathlib import Path
+from typing import Self, cast
 
 import polib
 from babel import Locale as BabelLocale
 from pydantic import BaseModel
 
 from mex.common.context import SingleSingletonStore
-
-if TYPE_CHECKING:
-    from pathlib import Path
+from mex.common.transform import camelcase_to_title
 
 LOCALE_SERVICE_STORE = SingleSingletonStore["LocaleService"]()
 
@@ -27,19 +26,6 @@ class MExLocale(BaseModel):
     def values(self) -> Iterable[str]:
         """Expose locale values to avoid reflex bug."""
         return self.model_dump().values()  # sigh, don't ask
-
-
-# TODO(ND): Move this to mex-common
-def camelcase_to_title(value: str) -> str:
-    """Convert a camelcase string into title-cased words splitted by space.
-
-    Args:
-        value: The camelcase string to convert.
-
-    Returns:
-        The converted string containing title-cased words splitted by space.
-    """
-    return re.sub(r"(?<!^)(?=[A-Z])", " ", value).title()
 
 
 class LocaleService: